ai-engine/requirements.txt
torch==2.5.0
transformers==4.45.0
optuna==3.6.0
prophet==1.2.4
scikit-learn==1.6.1
joblib==1.4.2
pandas==2.2.3
numpy==2.1.2
requests==2.32.3
boto3==1.35.0
minio==7.2.7
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
prometheus-client==0.21.0
pyjwt==2.9.0
